+++ Home/urls.py
@@ -9,7 +9,5 @@ urlpatterns = [
     path('logout/', auth_views.LogoutView.as_view(template_name="/"), name='logout_page'),
     path('signup', views.signup, name="sign_up_page"),
     path('profile', views.profile, name="profile_page"),
-    path('game', views.game, name="game"),
-    path('api/game', views.get_nplayer_game, name="get_game"),
     path('join', views.join, name="join_page")
 ]
