+++ Home/views.py
@@ -85,35 +85,6 @@ def profile(request):
     # else:
     #     return render(request, 'Home/login.html', {})
 
-def get_nplayer_game(request):
-    if not request.GET:
-        err = {
-            'status': 400,
-            'id': None,
-            'text': 'Please enter number of players',
-            'game': None
-        }
-        return JsonResponse(err)
-    nplayers = request.GET.get('nplayers', None)
-    min_games = Game.objects.filter(game_type__min_players__gte=nplayers)
-    filtered_games = Game.objects.filter(game_type__max_players__lte=nplayers)
-    rand_game = random.choice(filtered_games)
-    context = {
-        'status': 200,
-        'id': rand_game.id,
-        'text': rand_game.text,
-        'game': rand_game.game_type.get_game_type_display()
-    }
-    return JsonResponse(context)
-
-def game(request):
-    username = None
-    userGroup = None
-    # if request.user.is_authenticated:
-    #     username = request.user.username
-    #     userGroup = Group
-    return render(request, 'Home/game.html')
-
 @login_required(login_url='login/')
 def join(request):
     if request.method == 'POST':
